Restructure elementTree into encapsulated, somewhat object oriented, blocks
