+++ frontend/testsrc/MainPage/CoursesArea/Label.tsx
