+++ frontend/testsrc/MainPage/MobilePanel/index.tsx
